Re-write View logic, separate business logic
