+++ videochat udp/MainWindow.resx
